Refactor(lib-meow): update =meow--select= calls with additional argument
This change addresses the impact of commit e720a19 in the Meow repository, which aimed to make the placement of marks consistent for the INSERT state (see issue #675). The update ensures that entering and exiting the INSERT state places marks consistently, regardless of the 'meow-select-on-*' variable settings. This involves adding configuration options for 'meow-open-*' functions, where the variables now solely determine if the resulting mark should be active upon exiting the INSERT state.